new package added - parser with methods

+++ src/main/java/entity/Boots.java
@@ -6,7 +6,7 @@ public class Boots extends Product {
     int size;
     boolean isNaturalSkin;
 
-    public Boots(int id, String productName, double price, double weight, String color, int productCount, int size) {
+    public Boots(long id, String productName, double price, double weight, String color, int productCount, int size) {
         super(id, productName, price, weight, color, productCount);
         this.size = size;
